requirements.txt
flask
qdrant-client
sentence-transformers
datasets
pillow
tqdm
transformers
